libnm-core: Add missing type cast
Fixes testing on 32-bit arches:

/core/general/test_setting_compare_timestamp:
(./test-general:29331): GLib-GObject-WARNING **: g_object_set_valist: object class `NMSettingConnection' has no property named `$?\xff\xff\x89t$0\x89|$4\xe8\u001c\x98\xff\xff\x85\xc0tM\x83\xf8\xfft3\x8dT$(\xc7D$\u0008'
/bin/sh: line 5: 29331 Trace/breakpoint trap   ${dir}$tst
FAIL: test-general

Fixes: 093a3c88d03bf91ed65e80e45261451a50e974df

+++ libnm-core/tests/test-general.c
@@ -2259,11 +2259,11 @@ test_setting_compare_timestamp (void)
 	              NM_SETTING_CONNECTION_ID, "ignore timestamp connection",
 	              NM_SETTING_CONNECTION_UUID, "b047a198-0e0a-4f0e-a653-eea09bb35e40",
 	              NM_SETTING_CONNECTION_AUTOCONNECT, FALSE,
-	              NM_SETTING_CONNECTION_TIMESTAMP, 1234567890,
+	              NM_SETTING_CONNECTION_TIMESTAMP, (guint64) 1234567890,
 	              NULL);
 
 	new = nm_setting_duplicate (old);
-	g_object_set (new, NM_SETTING_CONNECTION_TIMESTAMP, 1416316539, NULL);
+	g_object_set (new, NM_SETTING_CONNECTION_TIMESTAMP, (guint64) 1416316539, NULL);
 
 	/* First make sure they are different */
 	success = nm_setting_compare (old, new, NM_SETTING_COMPARE_FLAG_EXACT);
